+++ src/components/DeleteStatisticsModal/DeleteStatistics/index.js
@@ -1,0 +1,1 @@
+export { default } from './DeleteStatistics';
