+++ src/components/DeleteStatisticsModal/index.js
@@ -1,0 +1,1 @@
+export { default } from './DeleteStatisticsModal';
